Move pack_up to utils
